% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlm-package.R
\docType{package}
\name{bmlm-package}
\alias{bmlm-package}
\alias{bmlm}
\title{bmlm: Easy estimation of Bayesian multilevel mediation models with Stan.}
\description{
See \url{https://mvuorre.github.io/bmlm/} for a short tutorial.
}
