% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmlm.R
\docType{package}
\name{bmlm}
\alias{bmlm}
\alias{bmlm-package}
\title{bmlm: An R package for estimating a multilevel mediation model with rstan.}
\description{
bmlm: An R package for estimating a multilevel mediation model with rstan.
}

