% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stancode}
\alias{get_stancode}
\title{Get Stan code for bmlm model}
\usage{
get_stancode()
}
\value{
A text string including the Stan code.
}
\description{
Prints Stan code for bmlm model that's used in calls to \code{bmlm::mlm()}
}
\examples{
x <- get_stancode()
cat(x)
print(x)
}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

