% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getpps}
\alias{getpps}
\title{Get posterior probability}
\usage{
getpps(x)
}
\arguments{
\item{x}{A vector of MCMC samples.}
}
\value{
Posterior probability.
}
\description{
Calculates the proportion of samples on the dominant side of a reference value.
}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

