% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isolate}
\alias{isolate}
\title{Create isolated within- (and optionally between-) person variables.}
\usage{
isolate(d = NULL, by = NULL, value = NULL, z = FALSE,
  which = "within")
}
\arguments{
\item{d}{A \code{data.frame}.}

\item{by}{A vector of values by which the data is clustered.
i.e. a vector of unique participant IDs.}

\item{value}{Names of columns to isolate. Multiple values can be given by
\code{value = c("var1", "var2", "var3")}}

\item{z}{Should the new values be standardized (defaults to FALSE).}

\item{which}{Which component to return. "within" (default) returns
within-person deviations only; "between" returns between-person means only;
"both" returns both.}
}
\value{
A \code{data.frame} with additional columns for the within- and
between-person variables. The new columns are labelled _cw for
centered-within and _cb for centered-between.
}
\description{
Creates variables that represent pure within- and between-person predictors.
}
\examples{
\dontrun{
# Create within-person deviations of \\code{x}.
mydata <- isolate(mydata, by = "id", value = "x")  # Overwrites mydata
}

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

