% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlm}
\alias{mlm}
\title{Bayesian parameter estimation for multilevel mediation models}
\usage{
mlm(d = NULL, id = "id", x = "x", m = "m", y = "y",
  prior_scale = NULL, tau_scale = NULL, intrcpt_scale = NULL, ...)
}
\arguments{
\item{d}{A \code{data.frame}.}

\item{id}{Column of participant IDs in \code{data}.}

\item{x}{Column of X values in \code{data}.}

\item{m}{Column of M values in \code{data}.}

\item{y}{Column of Y values in \code{data}.}

\item{prior_scale}{Prior standard deviation on regression coefficients.
See details.}

\item{tau_scale}{Prior scale on \code{tau} parameters. See details.}

\item{intrcpt_scale}{Prior standard deviation on intercepts.}

\item{...}{Other optional parameters passed to \code{rstan::stan()}.}
}
\value{
An object of S4 class stanfit, with all its available methods.
}
\description{
Easy Bayesian parameter estimation for multilevel mediation models.
}
\details{
Draw samples from the joint posterior distribution of a
multilevel mediation model using Stan.

\subsection{Priors}{

\code{prior_scale} inputs a standard deviation parameter to the prior
distributions for the regression parameters in a, b, and cp paths. Users are
recommended to adjust this to fit the scale of the data.
\code{tau_scale} inputs a scale parameter to Cauchy distributions on
the varying effects' standard deviation parameters.

}

Currently, we assume that the user inputs X, M, and Y as within-person
deviated variables (Bolger & Laurenceau, 2013, ch.9). As a result, the
regression intercepts will be zero.
}
\examples{
\dontrun{
## Run example from Bolger and Laurenceau (2013)
data(BLch9)
fit <- mlm(BLch9)
mlm_summary(fit)
}

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

