% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_plot.R
\name{mlm_plot}
\alias{mlm_plot}
\title{Plot multilevel mediation}
\usage{
mlm_plot(mod = NULL, xlab = "X", ylab = "Y", mlab = "M")
}
\arguments{
\item{mod}{A Stanfit model estimated with \code{bmlm::mlm()}.}

\item{xlab}{Label for X}

\item{ylab}{Label for Y}

\item{mlab}{Label for M}
}
\value{
A qgraph object.
}
\description{
Plots a path diagram from a multilevel mediation model.
}
\details{
Experimental. Plots a path diagram of the mediation model,
with estimated average parameter values and posterior probabilities.
}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

