% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_model}
\alias{print_model}
\title{Get Stan code for bmlm model}
\usage{
print_model()
}
\value{
A text string including the Stan code.
}
\description{
Prints Stan code for bmlm model that's used in calls to \code{bmlm::mlm()}
}
\examples{
x <- print_model()
cat(x)
print(x)
}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

