% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bmlm.R
\name{summary.bmlm}
\alias{summary.bmlm}
\title{summary() method for an bmlm object}
\usage{
\method{summary}{bmlm}(bmlm, r = 3)
}
\value{
Returns a table with regression results.
}
\description{
summary() method for an bmlm object
}
\examples{
data(coalgov)
m1 <- bmlm(Surv(govdur, earlyterm, govmaxdur) ~ 1 + mm(id(pid, gid), mmc(fdep), mmw(w ~ 1/n, constraint=1)) + majority + hm(id=cid, name=cname, type=RE, showFE=F),
          family="Weibull", monitor=T, data=coalgov)
summary(m1)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
