% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tab2doc}
\alias{tab2doc}
\title{Create a word document from a summary table.}
\usage{
tab2doc(d = NULL, name = NULL)
}
\arguments{
\item{d}{A \code{data.frame}.}

\item{name}{Name of file to create. Defaults to "Table.docx".}
}
\value{
Saves a word document in the current working directory.
}
\description{
Saves a Word document to the current working directory.
}
\details{
Requires the ReporteRs R package. Copy-pasting individual values
is error-prone. Use \code{tab2doc()} to create a word document containing
a summary table.
}
\examples{
\dontrun{
tab2doc(mlm_summary(fit), name = "Fit_summary")
}

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

