% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getpps}
\alias{getpps}
\title{Get posterior probability}
\usage{
getpps(x, ref_val = 0)
}
\arguments{
\item{x}{A vector of MCMC samples.}

\item{ref_val}{Reference value, i.e. if samples are greater/smaller than
this. (Defaults to 0.)}
}
\value{
Posterior probability.
}
\description{
Calculates the proportion of samples on the dominant side of a reference value.
}
\examples{
# Probability that standard normal deviates are greater than 1
getpps(rnorm(1000), ref_val = 1)

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

