% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlm}
\alias{mlm}
\title{Bayesian parameter estimation for multilevel mediation models}
\usage{
mlm(d = NULL, id = id, x = x, m = m, y = y, ...)
}
\arguments{
\item{d}{A \code{data.frame}.}

\item{id}{Column of participant IDs in \code{data}.}

\item{x}{Column of X values in \code{data}.}

\item{m}{Column of M values in \code{data}.}

\item{y}{Column of Y values in \code{data}.}

\item{...}{Other optional parameters passed to \code{rstan::stan()}.}
}
\value{
An object of S4 class stanfit, with all its available methods.
}
\description{
Easy Bayesian parameter estimation for multilevel mediation models.
}
\details{
Draw samples from the joint posterior distribution of a
multilevel mediation model using Stan.

By default, the procedure doesn't return all the estimated parameters,
such as covariances etc. To return all estimated parameters,
set \code{everything = TRUE}.
}
\examples{
\dontrun{
## Run example from Bolger and Laurenceau (2013)
data(BLch9)
fit <- mlm(BLch9)
print(fit)}

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

