% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{mlm}
\alias{mlm}
\title{Bayesian parameter estimation for multilevel mediation models}
\usage{
mlm(d = NULL, id = id, x = x, m = m, y = y, prior_scale = NULL,
  intrcpt_scale = NULL, no_priors = FALSE, ...)
}
\arguments{
\item{d}{A \code{data.frame}.}

\item{id}{Column of participant IDs in \code{data}.}

\item{x}{Column of X values in \code{data}.}

\item{m}{Column of M values in \code{data}.}

\item{y}{Column of Y values in \code{data}.}

\item{prior_scale}{Prior standard deviation on regression coefficients.
See details.}

\item{intrcpt_scale}{Prior standard deviation on regression intercepts.
See details}

\item{no_priors}{Estimates the model with no priors. Not recommended,
see details.}

\item{...}{Other optional parameters passed to \code{rstan::stan()}.}
}
\value{
An object of S4 class stanfit, with all its available methods.
}
\description{
Easy Bayesian parameter estimation for multilevel mediation models.
}
\details{
Draw samples from the joint posterior distribution of a
multilevel mediation model using Stan.

\code{mlm()} Allows the user to input standard deviation parameters for
the regression coefficients and intercepts. If no priors are provided,
these are set to 100 (unless no_priors = TRUE).

If the user sets no_priors = TRUE, all parameters are estimated with
uniform priors. This is generally not recommended, but in some cases
the default priors might not be suitable to the data. \code{mlm()} uses
weakly regularizing priors by default, which help avoid overfitting.

Currently, we assume that the user inputs X, M, and Y as within-person
deviated variables (Bolger & Laurenceau, 2013, ch.9). As a result, the
regression intercepts will be zero.
}
\examples{
\dontrun{
## Run example from Bolger and Laurenceau (2013)
data(BLch9)
fit <- mlm(BLch9)
mlm_summary(fit)
}

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

