% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_plot.R
\name{mlm_pars_plot}
\alias{mlm_pars_plot}
\title{Plot marginal posterior histograms}
\usage{
mlm_pars_plot(mod = NULL, type = "hist", color = "black", level = 0.91,
  pars = c("a", "b", "cp", "corrab", "ab", "c", "pme"))
}
\arguments{
\item{mod}{A Stanfit model estimated with \code{bmlm::mlm()}.}

\item{type}{Type of the plot, \code{hist} or \code{coefplot}.}

\item{color}{Color for plots.}

\item{level}{X level for Credible Intervals.}

\item{pars}{List of parameters to plot.}
}
\value{
A ggplot2 object.
}
\description{
Plot the model's estimated parameters as histograms or a coefficient plot.
}
\details{
Experimental. Plots a path diagram of the mediation model,
with estimated average parameter values and posterior probabilities.
}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

