% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm_plot.R
\name{mlm_path_plot}
\alias{mlm_path_plot}
\title{Plot multilevel mediation as a path diagram}
\usage{
mlm_path_plot(mod = NULL, xlab = "X", ylab = "Y", mlab = "M",
  border.width = 2, edge.label.cex = 1, edge.color = "black",
  fade = FALSE, level = 0.91, text = FALSE, template = FALSE, ...)
}
\arguments{
\item{mod}{A Stanfit model estimated with \code{bmlm::mlm()}.}

\item{xlab}{Label for X}

\item{ylab}{Label for Y}

\item{mlab}{Label for M}

\item{border.width}{Size of node borders (defaults to 2).}

\item{edge.label.cex}{Text size.}

\item{edge.color}{Color of the edges.}

\item{fade}{Should edges fade to white? (Defaults to FALSE.)}

\item{level}{"Confidence" level for credible intervals.}

\item{text}{Should additional parameter values be displayed?}

\item{template}{Should an empty template diagram be plotted?}

\item{...}{Other arguments passed on to \code{qgraph::qgraph()}.}
}
\value{
A qgraph object.
}
\description{
Plots a path diagram from a multilevel mediation model.
}
\details{
Experimental. Plots a path diagram of the mediation model,
with estimated average parameter values and credible intervals. Can also
be used to draw a template diagram of the mediation model.
}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

